fix: update test-utils import paths after directory restructuring
