+++ src/components/layout/page-sections/lister-section/rr-lister-section.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-lister-section.ts';
 
 describe('rr-lister-section', () => {
