+++ src/components/layout/page-sections/one-half-one-half-section/rr-one-half-one-half-section.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-one-half-one-half-section.ts';
 
 describe('rr-one-half-one-half-section', () => {
