+++ src/components/layout/page-sections/one-third-two-thirds-section/rr-one-third-two-thirds-section.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-one-third-two-thirds-section.ts';
 
 describe('rr-one-third-two-thirds-section', () => {
