+++ src/components/layout/page-sections/simple-section/rr-simple-section.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-simple-section.ts';
 
 describe('rr-simple-section', () => {
