+++ src/components/layout/split-views/horizontal-split-view/rr-horizontal-split-view.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-horizontal-split-view.ts';
 
 describe('rr-horizontal-split-view', () => {
