+++ src/components/layout/split-views/side-by-side-split-view/rr-side-by-side-split-view.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-side-by-side-split-view.ts';
 
 describe('rr-side-by-side-split-view', () => {
