+++ src/components/layout/split-views/split-view-pane/rr-split-view-pane.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-split-view-pane.ts';
 
 describe('rr-split-view-pane', () => {
