+++ src/components/layout/split-views/stacked-split-view/rr-stacked-split-view.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-stacked-split-view.ts';
 
 describe('rr-stacked-split-view', () => {
