+++ src/components/layout/split-views/vertical-split-view/rr-vertical-split-view.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-vertical-split-view.ts';
 
 describe('rr-vertical-split-view', () => {
