+++ src/components/lists-and-menus/cells/button-cell/rr-button-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-button-cell.ts';
 
 describe('rr-button-cell', () => {
