+++ src/components/lists-and-menus/cells/description-cell/rr-description-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-description-cell.ts';
 
 describe('rr-description-cell', () => {
