+++ src/components/lists-and-menus/cells/icon-cell/rr-icon-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-icon-cell.ts';
 
 describe('rr-icon-cell', () => {
