+++ src/components/lists-and-menus/cells/label-cell/rr-label-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-label-cell.ts';
 
 describe('rr-label-cell', () => {
