+++ src/components/lists-and-menus/cells/spacer-cell/rr-spacer-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-spacer-cell.ts';
 
 describe('rr-spacer-cell', () => {
