+++ src/components/lists-and-menus/cells/stepper-cell/rr-stepper-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-stepper-cell.ts';
 
 describe('rr-stepper-cell', () => {
