+++ src/components/lists-and-menus/cells/text-cell/rr-text-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-text-cell.ts';
 
 describe('rr-text-cell', () => {
