+++ src/components/lists-and-menus/cells/timeline-track-cell/rr-timeline-track-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-timeline-track-cell.ts';
 
 describe('rr-timeline-track-cell', () => {
