+++ src/components/lists-and-menus/cells/title-cell/rr-title-cell.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { fixture, cleanup, waitForUpdate } from '../../../test-utils.ts';
+import { fixture, cleanup, waitForUpdate } from '../../../../test-utils.ts';
 import './rr-title-cell.ts';
 
 describe('rr-title-cell', () => {
